Rename PDFMatrix file
