+++ lib/prawn-svg.rb
@@ -10,7 +10,7 @@ require 'prawn/svg/calculators/arc_to_bezier_curve'
 require 'prawn/svg/calculators/aspect_ratio'
 require 'prawn/svg/calculators/document_sizing'
 require 'prawn/svg/calculators/pixels'
-require 'prawn/svg/transform_utils'
+require 'prawn/svg/pdfmatrix'
 require 'prawn/svg/transform_parser'
 require 'prawn/svg/url_loader'
 require 'prawn/svg/loaders/data'
